fix: Consolidate API endpoints to stay under Vercel limit
- Merge 5 PM endpoints into single /api/pm handler with action parameter
- Merge 3 config endpoints into single /api/config handler
- Remove duplicate v2 endpoints (submit-service-v2, submit-garden-v2)
- Reduce total endpoints from 18 to 10 (under Vercel Hobby limit of 12)

Changes:
- /api/pm handles: create-project, get-projects, get-tasks, update-task, get-dashboard
- /api/config handles: get-settings, update-setting, get-history
- All functionality preserved with action-based routing
